Revert "Final creo #2"
This reverts commit b6cc0adc90d5df31a07ade42d6949e9a0cab439b.

+++ controllers/categoryController.js
@@ -2,6 +2,14 @@ const Category = require('../models/Category');
 
 exports.createCategory = async (req, res) => {
     try {
+
+        if (req.userRole !== "admin" && req.userRole !== "coordinador") {
+            return res.status(403).json({
+                success: false,
+                message: "Solo administradores o coordinadores pueden crear categoria"
+            });
+        }
+
         const { name, description } = req.body;
 
         //Validacion
@@ -106,6 +114,13 @@ exports.getCategoryById = async (req, res) => {
 
 exports.updateCategory = async (req, res) => {
     try {
+        if (req.userRole !== "admin" && req.userRole !== "coordinador") {
+            return res.status(403).json({
+                success: false,
+                message: "Solo administradores o coordinadores pueden actualizar"
+            });
+        }
+
         const { name, description } = req.body;
         const updateData = {};
 
@@ -157,6 +172,14 @@ exports.updateCategory = async (req, res) => {
 
 exports.deleteCategory = async (req, res) => {
     try {
+
+        if (req.userRole !== "admin") {
+            return res.status(403).json({
+                success: false,
+                message: 'No tienes permiso para esta accion'
+            })
+        }
+
         const deleteCategory = await Category.findByIdAndDelete(req.
             params.id);
         if (!deleteCategory) {
